refactor: timezones in its own file
